Separated the parameter factory from the request factory.
